reorganize ccc
