+++ src/main/scala/stuff/boolean.scala
@@ -15,7 +15,7 @@ case object booleans {
     λ { b: scala.Boolean => if(b) ⊤ else ⊥  }
 
   val toBoolean: Ω -> scala.Boolean =
-    either { λ { s: ∗ => true } & λ { s: ∗ => false } }
+    either { point(true) and point(false) }
 
   def eq[A]: A × A -> Ω =
     λ { as: A × A => left(as) == right(as) } >-> fromBoolean
@@ -25,4 +25,22 @@ case object booleans {
 
   val ∨ : Ω × Ω -> Ω =
     ???
+
+  // better a curried version?
+  def If[X,Y]: (X -> Ω) -> (((∗ -> Y) × (∗ -> Y)) -> (X -> Y)) =
+    λ { p => λ { tf => p >-> either(tf) } }
+
+  import scala.Int
+
+  val isEven: Int -> Ω =
+    λ { x => eq(x % 2 and 0) }
+
+  val msg =
+    point("Is Even!")
+
+  val other =
+    point("Is odd!")
+
+  val tellMe =
+    If(isEven)(msg and other)
 }
